fix: code optimization+ preview audio updated
